+++ server/voice_changer/RVC/inferencer/WebUIInferencerNono.py
@@ -46,5 +46,5 @@ class WebUIInferencerNono(Inferencer):
             return_length=return_length,
             formant_length=formant_length
         )
-        res = res[0][0, 0].float()
+        res = res[0][0, 0]
         return torch.clip(res, -1.0, 1.0)
